# src/dto/user.dto.ts
export interface UserInputDTO {
  username: string;
  password: string;
}

export interface UserInputPatchDTO {
  username?: string;
  password?: string;
}

export interface UserOutputDTO {
  id: number;
  username: string;
  password: string;
  role: string;
}